MOVE-302: DPA

+++ integrasjonspunkt/src/main/java/no/difi/meldingsutveksling/nextmove/NextMoveSender.java
@@ -1,17 +1,22 @@
 package no.difi.meldingsutveksling.nextmove;
 
 import lombok.extern.slf4j.Slf4j;
+import no.difi.meldingsutveksling.ServiceIdentifier;
 import no.difi.meldingsutveksling.nextmove.message.MessagePersister;
 import no.difi.meldingsutveksling.receipt.ConversationService;
 import no.difi.meldingsutveksling.receipt.GenericReceiptStatus;
 import no.difi.meldingsutveksling.receipt.MessageStatus;
+import no.difi.meldingsutveksling.serviceregistry.ServiceRegistryLookup;
+import no.difi.meldingsutveksling.serviceregistry.externalmodel.ServiceRecord;
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import javax.transaction.Transactional;
 import java.io.IOException;
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import static no.difi.meldingsutveksling.nextmove.ConversationDirection.OUTGOING;
 import static no.difi.meldingsutveksling.nextmove.logging.ConversationResourceMarkers.markerFrom;
@@ -24,6 +29,7 @@ public class NextMoveSender {
     private ConversationService conversationService;
     private MessagePersister messagePersister;
     private DirectionalConversationResourceRepository outRepo;
+    private ServiceRegistryLookup sr;
 
     @Autowired
     public NextMoveSender(ConversationStrategyFactory strategyFactory,
@@ -45,6 +51,21 @@ public class NextMoveSender {
             log.error(markerFrom(cr), errorStr);
             throw new NextMoveRuntimeException(errorStr);
         }
+
+        List<ServiceRecord> serviceRecords = sr.getServiceRecords(cr.getReceiverId());
+        Optional<ServiceRecord> serviceRecord = serviceRecords.stream()
+                .filter(r -> cr.getServiceIdentifier() == r.getServiceIdentifier())
+                .findFirst();
+        if (!serviceRecord.isPresent()) {
+            List<ServiceIdentifier> acceptableTypes = serviceRecords.stream()
+                    .map(ServiceRecord::getServiceIdentifier)
+                    .collect(Collectors.toList());
+            String errorStr = String.format("Message is of type '%s', but receiver '%s' accepts types '%s'.",
+                    cr.getServiceIdentifier(), cr.getReceiverId(), acceptableTypes);
+            log.error(markerFrom(cr), errorStr);
+            throw new NextMoveException(errorStr);
+        }
+
         strategy.get().send(cr);
         conversationService.registerStatus(cr.getConversationId(), MessageStatus.of(GenericReceiptStatus.SENDT));
         outRepo.delete(cr);
